requirements.txt
pyserial==2.6
wsgiref==0.1.2
